+++ client/pages/404.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+import ParagraphText from "+/lib/components/ParagraphText";
+
+const NotFoundPage = () => (
+    // TODO: Flesh this out
+    <>
+        <h1>Oh snap!</h1>
+        <ParagraphText
+            text="This page does not exist, or maybe our database is down. Who knows?
+            Not you. Hahahaha"
+        />
+    </>
+);
+
+export default NotFoundPage;