+++ client/pages/index.tsx
@@ -1,47 +1,31 @@
-import React, { useContext, useEffect, useState } from "react";
-import { useRouter } from "next/router";
+import React, { useContext } from "react";
 
 import CrowdventureButton from "+/lib/components/CrowdventureButton";
 import { ModalContext } from "+/lib/modal";
 import NodeViewer from "+/lib/nodes/NodeViewer";
 import CreateNodeModal from "+/lib/nodes/CreateNodeModal";
-import { UnsafeModeContext } from "+/lib/unsafeMode";
 
 import { type Node } from "@/types/models";
+import apiClient from "+/lib/apiClient";
+import { useSafeGuardedNodes } from "+/lib/specialHooks";
+
+const getFeaturedNodes = async (unsafeMode: boolean) => {
+    const response = await apiClient.provide("get", "/node/featuredNodes", {
+        allowHidden: String(unsafeMode),
+    });
+    if (response.status === "error") throw new Error(response.error.message);
+    return response.data.nodes;
+};
 
 const HomePage = ({
-    topNodes: initTopNodes,
-    recentNodes: initRecentNodes,
+    featuredNodes: initFeaturedNodes,
 }: {
-    readonly topNodes: Node[];
-    readonly recentNodes: Node[];
+    readonly featuredNodes: Node[];
 }) => {
-    const { unsafeMode } = useContext(UnsafeModeContext);
     const { openModal } = useContext(ModalContext);
-    const [page, setPage] = useState(1);
-    const [topNodes, setTopNodes] = useState(initTopNodes);
-    const [recentNodes, setRecentNodes] = useState(initRecentNodes);
-
-    const router = useRouter();
-
-    useEffect(() => {
-        queryCall("featuredNodes", NODE_PREVIEW_GQL, {
-            allowHidden: unsafeMode,
-        }).then(setTopNodes);
-    }, [unsafeMode]);
-
-    useEffect(() => {
-        // queryCall("recentlyUpdatedNodes", NODE_PREVIEW_GQL, {
-        //     pageNum: page,
-        //     allowHidden: unsafeMode,
-        // }).then(setRecentNodes);
-    }, [unsafeMode, page]);
-
-    const goToRandomNode = () => {
-        queryCall("randomNode", { ID: 0 }, { allowHidden: unsafeMode }).then(
-            (randomNode) => router.push(`/node/${randomNode.ID}`)
-        );
-    };
+    const featuredNodes = useSafeGuardedNodes(initFeaturedNodes, () =>
+        getFeaturedNodes(true)
+    );
 
     return (
         <>
@@ -55,55 +39,15 @@ const HomePage = ({
             <hr />
 
             <h3>Featured Stories:</h3>
-            <NodeViewer nodes={topNodes} />
-
-            <hr />
-
-            {/* <h3>Recently added or updated:</h3>
-            <NodeViewer nodes={recentNodes} />
-
-            <hr />
-
-            <CrowdventureButton
-                // This button needs to be formatted off to the side
-                onClick={() => {
-                    setPage(page + 1);
-                }}
-            >
-                Next &gt;
-            </CrowdventureButton> */}
-
-            <CrowdventureButton
-                onClick={() => {
-                    goToRandomNode();
-                }}
-            >
-                Play a random adventure!
-            </CrowdventureButton>
+            <NodeViewer nodes={featuredNodes} />
         </>
     );
 };
 
-export const getStaticProps = async () => {
-    const unsafeMode = false;
-    const page = 1;
-
-    return {
-        props: {
-            topNodes: await queryCall("featuredNodes", NODE_PREVIEW_GQL, {
-                allowHidden: unsafeMode,
-            }),
-            // recentNodes: await queryCall(
-            //     "recentlyUpdatedNodes",
-            //     NODE_PREVIEW_GQL,
-            //     {
-            //         pageNum: page,
-            //         allowHidden: unsafeMode,
-            //     }
-            // ),
-            recentNodes: [],
-        },
-    };
-};
+export const getStaticProps = async () => ({
+    props: {
+        featuredNodes: await getFeaturedNodes(false),
+    },
+});
 
 export default HomePage;